Chore: Refactor Modal Folder Structure & Lint Rule
