+++ src/Compoents/Modals/Deck/homeModal.connector.js
@@ -4,7 +4,7 @@ import { cancelEditModalHome } from 'src/redux/actions';
 import { useNavigation } from '@react-navigation/native';
 import { handleMainButton, handleCancelButton, getTitle } from './homeModal.helper';
 
-import NativeModal from './Modal.react';
+import NativeModal from './HomeModal.react';
 
 const HomeModal = () => {
   const dispatch = useDispatch();
