+++ src/Compoents/Modals/index.js
@@ -1,3 +1,2 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as QuestionModal } from './QuestionModal.connector';
-export { default as HomeModal } from './homeModal.connector';
+export { default as QuestionModal } from './Question/QuestionModal.connector';
+export { default as HomeModal } from './Deck/homeModal.connector';
