+++ src/utils/helper.js
@@ -1,6 +1,5 @@
 import uuid from './uuid';
 
-// eslint-disable-next-line import/prefer-default-export
 export const formatCreateDeck = (title) => ({
   [title]: {
     title,
