+++ src/utils/index.js
@@ -1,4 +1,3 @@
-/* eslint-disable import/prefer-default-export */
 export { darkBlue, red, greyBlue } from './colors';
 
 export { default as uuid } from './uuid';
